# perf: massive performance improvements for project / sources manager