--- a/src/ape/managers/project.py
+++ b/src/ape/managers/project.py
@@ -1,6 +1,7 @@
 import json
 import random
 import shutil
+import time
 from collections.abc import Callable, Iterable, Iterator
 from contextlib import contextmanager
 from functools import cached_property, singledispatchmethod