--- a/tests/performance/test_contracts.py
+++ b/tests/performance/test_contracts.py
@@ -12,4 +12,4 @@ def test_compile_performance(benchmark, project_with_contracts):
     # Get the median of the measured times
     stats = benchmark.stats
     median = stats.get("median")
-    assert median < 0.01
+    assert median < 0.0002