### What I did

Before:

```
In [1]: time project.get_contract("Flat")
CPU times: user 18.3 ms, sys: 1.52 ms, total: 19.8 ms
Wall time: 18.6 ms
Out[1]: <Flat>

In [2]: time project.get_contract("Flat")
CPU times: user 3.14 ms, sys: 1.13 ms, total: 4.27 ms
Wall time: 3.34 ms
Out[2]: <Flat>
```

After:

```
In [1]: time project.get_contract("Flat")
CPU times: user 19.4 ms, sys: 1.61 ms, total: 21 ms
Wall time: 19.7 ms
Out[1]: <Flat>

In [2]: time project.get_contract("Flat")
CPU times: user 50 µs, sys: 0 ns, total: 50 µs
Wall time: 55.1 µs
Out[2]: <Flat>
```

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated