@antazoey 2024-07-24
- [ ] Add performance test